feat(pm): restrict team CRUD to super-admins

+++ src/Http/Controllers/Workspace/TeamController.php
@@ -12,6 +12,7 @@ use RayzenAI\ProjectManagement\Http\Resources\MemberResource;
 use RayzenAI\ProjectManagement\Http\Resources\TeamResource;
 use RayzenAI\ProjectManagement\Models\Member;
 use RayzenAI\ProjectManagement\Models\Team;
+use RayzenAI\ProjectManagement\Support\WorkspaceAccess;
 
 class TeamController extends Controller
 {
@@ -48,6 +49,8 @@ class TeamController extends Controller
 
     public function destroy(Team $team): RedirectResponse
     {
+        abort_unless(WorkspaceAccess::isSuperAdmin(request()->user()), 403);
+
         $team->delete();
 
         return back()->with('workspace_flash', ['success' => true, 'message' => 'Team deleted.']);
